requirements.txt
flask
python-dotenv
requests
yfinance
pandas
numpy
